solvebio/contrib/streamlit/dev-requirements.txt
httpx-oauth==0.3.7
streamlit==1.1.0
pandas==1.3.4
python-dotenv==0.19.1